+++ test/dimse/scu/n_delete_test.exs
@@ -4,6 +4,27 @@ defmodule Dimse.Scu.NDeleteTest do
   alias Dimse.Scu.NDelete
   alias Dimse.Command.Fields
 
+  defmodule FakeAssociation do
+    use GenServer
+
+    def start_link(response), do: GenServer.start_link(__MODULE__, response)
+
+    @impl true
+    def init(response), do: {:ok, response}
+
+    @impl true
+    def handle_call({:dimse_request, _command_set, _data}, _from, response) do
+      {:reply, response, response}
+    end
+  end
+
+  describe "send/4 transport error" do
+    test "propagates transport-level error from association" do
+      {:ok, assoc} = FakeAssociation.start_link({:error, :timeout})
+      assert {:error, :timeout} = NDelete.send(assoc, "1.2.3", "4.5.6")
+    end
+  end
+
   describe "build_command_set/3" do
     test "uses RequestedSOPClassUID (0000,0003)" do
       cmd = NDelete.build_command_set("1.2.3", "1.2.3.4", 1)